Initial Wipe line and Retraction Again
The g-code from Cura for pre and post build were copied over in order to ensure that the print head was clean before the print started and so that the filament wa retracted slightly after the print was finished.

Also, the previous version had stringing when jumping. The retraction length was increased from 2 to 6 and the retraction speed was increased from 40 to 80mm/s, which are the values that were being used in Cura. This has solved the problem